inference: store the KV cache as f16 by default
The flags have taken f16 on both sides since the cache types landed, but the default stayed f32 pending a decision, which the user took today.
f16 is what the reference runtime stores by default, so a comparison between the two needs no flags to be like-for-like.
It costs nothing the correctness gate can see: the 8B Q8_0 excerpt reads a perplexity delta of 0.001254 against the Hugging Face reference where f32 reads 0.001374, both inside a 0.01 bound.
It halves the cache, so Qwen3-8B runs a 16k context that f32 cannot allocate on a 16 GB card, and since a decode step reads the whole cache a 512-token generation on Qwen3-0.6B gains 6 percent.

The three components that compare exact f32 arithmetic against independently generated fixtures now ask for f32 cache sides themselves, so they keep testing what they tested before, and an explicit --cache-type still overrides them.
All three suites pass at the new default: the CPU suite, the CPU suite pinned to f32, and the device suite.

+++ src/inference/sampler.hpp
@@ -33,8 +33,8 @@ struct GenParams {
     int threads = 0;        // 0 = auto; used for generation (decode)
     int threads_batch = 0;  // 0 = same as threads; used for prefill
     int ubatch = 0;         // physical batch for prefill (0 = default 512)
-    std::string cache_type_k = "f32";   // KV cache storage per side: f32 or f16
-    std::string cache_type_v = "f32";
+    std::string cache_type_k = "f16";   // KV cache storage per side: f16 or f32
+    std::string cache_type_v = "f16";
     int kv_tokens = 0;      // the KV pool's total token budget (0 = the model context)
     std::string device = "cpu"; // backend: cpu, or vulkan:N when built with it
     float penalty = 1.0f;   // repetition penalty (>= 1)
